Code refactoring
This commit,
- Adds checks before casts
- Compares num_threads and thread_limit and chooses the loweset one
- Moves some code to a different function for readability

With the current implementation, target's thread_limit is
honored iff argument to thread_limit is a constant expr.
It also fails to affect an enclosed directive's num_threads clause
if the clause's argument is anything other than a constant expr.

TODO: support general expression arguments to thread_limit and
      num_threads