Initial support for general expressions
This commit adds support for general expressions as argument for
target thread_limit(). But the functionality of choosing the lower
limit when there is a conflicting num_threads() clause inside the
target thread_limit region is removed as it can't be done at parse
phase anymore.

TODO: support the cases where there is a conflicting num_threads
      clause in enclosed parallel directives.